FS-28 Extra nullchecks

+++ src/main/java/org/fides/server/files/FileManager.java
@@ -59,20 +59,25 @@ public final class FileManager {
 		if (temporary) {
 			location += ".tmp";
 		}
-		File newFile = new File(properties.getDataDir(), location);
+		String dataDir = PropertiesManager.getInstance().getDataDir();
+		if (StringUtils.isNotEmpty(dataDir)) {
+			File newFile = new File(dataDir, location);
 
-		// Check if the filename is unique, there's a maximum number of attempts to prevent an overflow
-		try {
-			int uniqueAttempts = 0;
-			while (!newFile.createNewFile() && ++uniqueAttempts <= MAXUNIQUENAMEATTEMPTS) {
-				location = UUID.randomUUID().toString();
-				if (temporary) {
-					location += ".tmp";
+			// Check if the filename is unique, there's a maximum number of attempts to prevent an overflow
+			try {
+				int uniqueAttempts = 0;
+				while (!newFile.createNewFile() && ++uniqueAttempts <= MAXUNIQUENAMEATTEMPTS) {
+					location = UUID.randomUUID().toString();
+					if (temporary) {
+						location += ".tmp";
+					}
+					newFile = new File(properties.getDataDir(), location);
 				}
-				newFile = new File(properties.getDataDir(), location);
+			} catch (IOException e) {
+				log.error(e);
+				location = null;
 			}
-		} catch (IOException e) {
-			log.error(e);
+		} else {
 			location = null;
 		}
 		// Return the location of the generated file
@@ -158,8 +163,12 @@ public final class FileManager {
 	 * @return If the file was successfully deleted. Returns false if the file doesn't exist.
 	 */
 	public static boolean removeFile(String location) {
-		File file = new File(PropertiesManager.getInstance().getDataDir(), location);
-		return file.delete();
+		String dataDir = PropertiesManager.getInstance().getDataDir();
+		if (StringUtils.isNotEmpty(dataDir) && StringUtils.isNotEmpty(location)) {
+			File file = new File(dataDir, location);
+			return file.delete();
+		}
+		return false;
 	}
 
 	/**
